+++ ci/build-init.groovy
@@ -27,6 +27,12 @@ pipeline {
         PYTHON_VERSION = sh(script: 'cat .python-version', returnStdout: true).trim()
 
         DOCKER_PROD_IMAGE = 'aio4/dev/services/oms/oms-sensemaking'
+
+        IMAGE_NAME="dpaas/ubi8-ccp"
+        IMAGE_VERSION="8.10-1262"
+
+        POSTGRES_REPOSITORY = "https://artifactory.code.dodiis.mil/artifactory/postgres-remote-cache"
+        EPEL_REPOSITORY = "https://artifactory.code.dodiis.mil/artifactory/epel-remote-cache"
     }
 
     stages {
@@ -37,7 +43,7 @@ pipeline {
                     filename 'ci/Dockerfile.jenkins'
                     registryUrl 'https://${artDockerUrl}'
                     registryCredentialsId env.SERVICE_ACCOUNT_ID
-                    additionalBuildArgs '--build-arg BASE_IMAGE=${artDockerUrl}/python:${PYTHON_VERSION}-slim'
+                    additionalBuildArgs '--build-arg BASE_IMAGE=${artDockerUrl}/{IMAGE_NAME}:${IMAGE_VERSION}'
                     args '''
                         -e HOME=/tmp \
                         -v /etc/pki/ca-trust/extracted/pem/tls-ca-bundle.pem:/etc/ssl/certs/ca-certificates.crt
@@ -107,10 +113,15 @@ pipeline {
                             docker build \
                                 -t ${artDockerUrl}/${DOCKER_PROD_IMAGE}:${APP_VERSION%%+*} \
                                 -t ${artDockerUrl}/${DOCKER_PROD_IMAGE}:latest \
+                                --build-arg NAMESPACE=${artDockerUrl} \
+                                --build-arg IMAGE_NAME=${IMAGE_NAME} \
+                                --build-arg IMAGE_VERSION=${IMAGE_VERSION} \
                                 --build-arg APP_VERSION=${APP_VERSION} \
                                 --build-arg APP_DATE=$(date -u +'%Y-%m-%dT%H:%M:%SZ') \
                                 --build-arg VCS_REF=$(git rev-parse HEAD) \
                                 --build-arg PIP_INDEX_URL=${artUrl}/api/pypi/pypi/simple \
+                                --build-arg POSTGRES_REPOSITORY=${POSTGRES_REPOSITORY} \
+                                --build-arg EPEL_REPOSITORY=${EPEL_REPOSITORY} \
                                 --secret id=mynetrc,src=.netrc \
                                 --secret id=cacert,src=/etc/pki/ca-trust/extracted/pem/tls-ca-bundle.pem \
                                 .